+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - adds pick function
 - memoize function
+- Maybe type and maybe constructor
+- mapMaybe function
+- Either type and its constructors: left, right, jsonError,and error
+- Either helpers: partition, tryCatch, and tryCatchAsync
 
 ### Changed
 - flatten function flattens by one level instead of Infinity